Move parser into logic/
