+++ src/main/java/chopchop/logic/parser/CommandArguments.java
@@ -1,6 +1,6 @@
 // CommandArguments.java
 
-package chopchop.parser;
+package chopchop.logic.parser;
 
 import java.util.List;
 import java.util.Optional;
