+++ src/test/java/chopchop/logic/parser/CommandParserTest.java
@@ -1,6 +1,6 @@
 // CommandParserTest.java
 
-package chopchop.parser;
+package chopchop.logic.parser;
 
 import java.util.Map;
